move helper functions in seperate folder and split to logical files
