+++ plugins/MySQL/boswatch.sql
@@ -1,3 +1,6 @@
+-- MySQL Database Structure for the BOSWatch MySQL Plugin
+-- @author: Bastian Schroll
+
 -- phpMyAdmin SQL Dump
 -- version 3.4.11.1deb2+deb7u1
 -- http://www.phpmyadmin.net
